Refactor HowYouCanHelp to standard form
